Allow http crate to be generic
