+++ src/webhooks/models/mod.rs
@@ -1,6 +1,5 @@
 mod docker;
 mod github;
-pub mod validators;
 
 pub use docker::Docker;
 pub use github::Github;
