+++ src/webhooks/validators.rs
@@ -1,4 +1,4 @@
-use crate::http::errors::AuthorizationError;
+use crate::http::AuthorizationError;
 use ring::hmac;
 use warp::{reject, Rejection};
